+++ 2FSemesterProjekt2024/Areas/Identity/Pages/Account/Register.cshtml.cs
@@ -32,18 +32,16 @@ namespace _2FSemesterProjekt2024.Areas.Identity.Pages.Account
         private readonly ILogger<RegisterModel> _logger;
         private readonly IEmailSender _emailSender;
         private readonly RoleManager<IdentityRole> _roleManager;
-        //Fordi det er strings rollerne er gemt i
         public List<string> AvailableRoles { get; set; }
 
-
         public RegisterModel(
-             UserManager<ApplicationUser> userManager,
-             IUserStore<ApplicationUser> userStore,
-             SignInManager<ApplicationUser> signInManager,
-             ILogger<RegisterModel> logger,
-             IEmailSender emailSender,
-             RoleManager<IdentityRole> roleManager) // Inject RoleManager
-{
+            UserManager<ApplicationUser> userManager,
+            IUserStore<ApplicationUser> userStore,
+            SignInManager<ApplicationUser> signInManager,
+            ILogger<RegisterModel> logger,
+            IEmailSender emailSender,
+            RoleManager<IdentityRole> roleManager)
+        {
             _userManager = userManager;
             _userStore = userStore;
             _emailStore = GetEmailStore();
@@ -51,9 +49,7 @@ namespace _2FSemesterProjekt2024.Areas.Identity.Pages.Account
             _logger = logger;
             _emailSender = emailSender;
             _roleManager = roleManager;
-           
         }
-
 
         /// <summary>
         ///     This API supports the ASP.NET Core Identity default UI infrastructure and is not intended to be used
@@ -108,22 +104,17 @@ namespace _2FSemesterProjekt2024.Areas.Identity.Pages.Account
             [Compare("Password", ErrorMessage = "The password and confirmation password do not match.")]
             public string ConfirmPassword { get; set; }
 
-
-            
             [Display(Name = "Select Roles")]
             public List<string> SelectedRoles { get; set; } = new List<string>();
-
-       
         }
 
 
-        public async Task OnGetAsync(string returnUrl = null) {
+        public async Task OnGetAsync(string returnUrl = null)
+        {
             ReturnUrl = returnUrl;
             ExternalLogins = (await _signInManager.GetExternalAuthenticationSchemesAsync()).ToList();
-            //VI konverter vores data til en liste så den kan læses
             AvailableRoles = await _roleManager.Roles.Select(r => r.Name).ToListAsync();
         }
-
 
         public async Task<IActionResult> OnPostAsync(string returnUrl = null)
         {
@@ -141,16 +132,13 @@ namespace _2FSemesterProjekt2024.Areas.Identity.Pages.Account
                 {
                     _logger.LogInformation("User created a new account with password.");
 
-                    // Here angiver vi rollerne ved at benytte nogle givet funktioner fra Identity pakken
                     foreach (var role in Input.SelectedRoles)
                     {
-                        if (await _roleManager.RoleExistsAsync(role))
-
+                        if(await _roleManager.RoleExistsAsync(role))
                         {
                             await _userManager.AddToRoleAsync(user, role);
                         }
                     }
-
 
                     var userId = await _userManager.GetUserIdAsync(user);
                     var code = await _userManager.GenerateEmailConfirmationTokenAsync(user);
@@ -180,7 +168,7 @@ namespace _2FSemesterProjekt2024.Areas.Identity.Pages.Account
                 }
             }
 
-            
+            // If we got this far, something failed, redisplay form
             return Page();
         }
