+++ 2FSemesterProjekt2024/Models/Booking.cs
@@ -1,37 +1,31 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
+﻿using Microsoft.AspNetCore.Identity;
 using System.ComponentModel.DataAnnotations.Schema;
-using Microsoft.EntityFrameworkCore;
+using System.ComponentModel.DataAnnotations;
 
-namespace _2FSemesterProjekt2024.Models;
-
-[Table("Booking")]
-public partial class Booking
+namespace _2FSemesterProjekt2024.Models
 {
-    [Key]
-    public int BookingId { get; set; }
 
-    public int? DriverId { get; set; }
+    [Table("Booking")]
+    public partial class Booking
+    {
+        [Key]
+        public int BookingId { get; set; }
 
-    public int? PassengerId { get; set; }
+        public int? DriverId { get; set; }
 
-    [StringLength(200)]
-    public string PickupLocation { get; set; } = null!;
+        public int? PassengerId { get; set; }
 
-    [StringLength(200)]
-    public string DropoffLocation { get; set; } = null!;
-    
-    public int Seats {  get; set; }
+        [StringLength(200)]
+        public string PickupLocation { get; set; } = null!;
 
-    [Column(TypeName = "datetime")]
-    public DateTime BookingTime { get; set; }
+        [StringLength(200)]
+        public string DropoffLocation { get; set; } = null!;
 
-    [ForeignKey("DriverId")]
-    [InverseProperty("Bookings")]
-    public virtual Driver? Driver { get; set; }
+        public int Seats { get; set; }
 
-    [ForeignKey("PassengerId")]
-    [InverseProperty("Bookings")]
-    public virtual Passenger? Passenger { get; set; }
+        [Column(TypeName = "datetime")]
+        public DateTime BookingTime { get; set; }
+
+     
+    }
 }
