--- 2FSemesterProjekt2024/Pages/Bookings/Create.cshtml
@@ -1,46 +1,0 @@
-﻿@page
-@model _2FSemesterProjekt2024.Pages.Bookings.CreateModel
-@{
-    ViewData["Title"] = "Create Booking";
-}
-
-<h1>Create Booking</h1>
-<div class="row">
-    <div class="col-md-4">
-        <form method="post">
-            <div asp-validation-summary="  ModelOnly  " class="text-danger">
-            </div>
-
-            <div class="form-group">
-                <label asp-for="@Model.Booking.PickupLocation" class="control-label"></label>
-                <input asp-for="@Model.Booking.PickupLocation" class="form-control" />
-                <span asp-validation-for="@Model.Booking.PickupLocation" class="text-danger"></span>
-            </div>
-            <div class="form-group">
-                <label asp-for="@Model.Booking.DropoffLocation" class="control-label"></label>
-                <input asp-for="@Model.Booking.DropoffLocation" class="form-control" />
-                <span asp-validation-for="@Model.Booking.DropoffLocation" class="text-danger"></span>
-            </div>
-
-            <div class="form-group">
-                <label asp-for="@Model.Booking.BookingTime" class="control-label"></label>
-                <input asp-for="@Model.Booking.BookingTime" class="form-control" />
-                <span asp-validation-for="@Model.Booking.BookingTime" class="text-danger"></span>
-            </div>
-
-            <div class="form-group">
-                <label asp-for="@Model.Booking.Seats" class="control-label"></label>
-                <input asp-for="@Model.Booking.Seats" class="form-control" />
-                <span asp-validation-for="@Model.Booking.Seats" class="text-danger"></span>
-            </div>
-
-            
-
-            <div class="form-group">
-                <input type="submit" value="Create" class="btn btn-primary" />
-                @*                   <a input asp-page="/Index" type="submit" value="Create" class="btn btn-primary">Create Profile</a>
-                *@
-            </div>
-        </form>
-    </div>
-</div>
