--- 2FSemesterProjekt2024/Pages/Bookings/Delete.cshtml
@@ -1,62 +1,0 @@
-﻿@page
-@model _2FSemesterProjekt2024.Pages.Bookings.DeleteModel
-@{
-
-}
-
-<h1>Delete Booking</h1>
-<h3>Are you sure you want to delete this?</h3>
-<div>
-    <form method="post">
-        <dl class="row">
-            <dt class="col-sm-2">
-                @Html.DisplayNameFor(model => model.Booking.BookingId)
-            </dt>
-            <dd class="col-sm-10">
-                @Html.DisplayFor(model => model.Booking.BookingId)
-            </dd>
-            <dt class="col-sm-2">
-                @Html.DisplayNameFor(model => model.Booking.Driver.DriverName) @* if these lines dont wrok, delete them *@
-            </dt>
-            <dd class="col-sm-10">
-                @Html.DisplayFor(model => model.Booking.Driver.DriverName) @* if these lines dont wrok, delete them *@
-            </dd>
-            <dt class="col-sm-2">
-                @Html.DisplayNameFor(model => model.Booking.Passenger.PassengerName) @* if these lines dont wrok, delete them *@
-            </dt>
-            <dd class="col-sm-10">
-                @Html.DisplayFor(model => model.Booking.Passenger.PassengerName) @* if these lines dont wrok, delete them *@
-            </dd>
-            <dt class="col-sm-2">
-                @Html.DisplayNameFor(model => model.Booking.PickupLocation)
-            </dt>
-            <dd class="col-sm-10">
-                @Html.DisplayFor(model => model.Booking.PickupLocation)
-            </dd>
-            <dt class="col-sm-2">
-                @Html.DisplayNameFor(model => model.Booking.DropoffLocation)
-            </dt>
-            <dd class="col-sm-10">
-                @Html.DisplayFor(model => model.Booking.DropoffLocation)
-            </dd>
-            <dt class="col-sm-2">
-                @Html.DisplayNameFor(model => model.Booking.BookingTime)
-            </dt>
-            <dd class="col-sm-10">
-                @Html.DisplayFor(model => model.Booking.BookingTime)
-            </dd>
-            <dt class="col-sm-2">
-                @Html.DisplayNameFor(model => model.Booking.Seats)
-            </dt>
-            <dd class="col-sm-10">
-                @Html.DisplayFor(model => model.Booking.Seats)
-            </dd>
-            
-
-        </dl>
-        <input type="submit" value="Delete" class="btn btn-danger" />
-    </form>
-    <div>
-        <a asp-page="GetDrivers">Back to List</a>
-    </div>
-</div>
