--- 2FSemesterProjekt2024/Pages/Bookings/Getbooking.cshtml
@@ -1,45 +1,0 @@
-﻿@page
-@model _2FSemesterProjekt2024.Pages.Bookings.GetBookingModel
-@{
-    ViewData["Title"] = "GetBooking";
-}
-
-<h3 style="color:red"> List of bookings</h3>
-<p>
-    <a asp-page="/Bookings/Create" asp-route-bid="BookingId">Create Booking</a>
-</p>
-<table class="table">
-    <thead>
-        <tr>
-            <th> PickupLocation </th>
-            <th> DropoffLocation </th>
-            <th> BookingTime </th>
-            <th> Seats </th>
-        </tr>
-    </thead>
-    <tbody>
-
-        @if (Model != null)
-        {
-            @foreach (var item in Model.Bookings)
-            {
-                <tr>
-                    <td>  @item.PickupLocation </td>
-                    <td>  @item.DropoffLocation  </td>
-                    <td>  @item.BookingTime  </td>
-                    <td>  @item.Seats  </td>
-                      
-                    <td>
-                       <p>
-
-                            <a asp-page="/Bookings/Delete" asp-route-bid="@item.BookingId">Delete</a>|
-                            <a asp-page="./Update" asp-route-bid="@item.BookingId">Edit</a>|
-                            
-                           
-                       </p>
-                    </td>
-                </tr>
-            }
-        }
-</table>
-
