--- 2FSemesterProjekt2024/Pages/Bookings/Update.cshtml
@@ -1,43 +1,0 @@
-﻿@page
-@model _2FSemesterProjekt2024.Pages.Bookings.UpdateModel
-@{
-    ViewData["Title"] = "UpdateBooking";
-}
-
-<h1>Update Booking</h1>
-
-<form method="post">
-    <div class="form-group">
-        <input type="hidden" asp-for="Booking.BookingId" />
-    </div>
-
-    <div class="form-group">
-        <label asp-for="@Model.Booking.PickupLocation" class="control-label"></label>
-        <input asp-for="@Model.Booking.PickupLocation" class="form-control" />
-        <span asp-validation-for="@Model.Booking.PickupLocation" class="text-danger"></span>
-    </div>
-    <div class="form-group">
-        <label asp-for="@Model.Booking.DropoffLocation" class="control-label"></label>
-        <input asp-for="@Model.Booking.DropoffLocation" class="form-control" />
-        <span asp-validation-for="@Model.Booking.DropoffLocation" class="text-danger"></span>
-    </div>
-
-    <div class="form-group">
-        <label asp-for="@Model.Booking.BookingTime" class="control-label"></label>
-        <input asp-for="@Model.Booking.BookingTime" class="form-control" />
-        <span asp-validation-for="@Model.Booking.BookingTime" class="text-danger"></span>
-    </div>
-
-    <div class="form-group">
-        <label asp-for="@Model.Booking.Seats" class="control-label"></label>
-        <input asp-for="@Model.Booking.Seats" class="form-control" />
-        <span asp-validation-for="@Model.Booking.Seats" class="text-danger"></span>
-    </div>
-    <div class="form-group">
-        <input type="submit" value="Save" class="btn btn-primary" />
-    </div>
-</form>
-
-<div>
-    <a asp-page="GetDriver">Back to List</a>
-</div>
