--- 2FSemesterProjekt2024/Pages/Drivers/Create.cshtml
@@ -1,55 +1,0 @@
-﻿@page
-@model _2FSemesterProjekt2024.Pages.Drivers.CreateModel
-@{
-    ViewData["Title"] = "Create";
-}
-
-<h1>Create</h1>
-<div class="row">
-    <div class="col-md-4">
-        <form method="post">
-            <div asp-validation-summary="  ModelOnly  " class="text-danger">
-            </div>
-
-            <div class="form-group">
-            <label asp-for="@Model.Driver.DriverName" class="control-label"></label>
-            <input asp-for="@Model.Driver.DriverName" class="form-control" />
-            <span asp-validation-for="@Model.Driver.DriverName" class="text-danger"></span>
-            </div>
-            <div class="form-group">
-                <label asp-for="@Model.Driver.Email" class="control-label"></label>
-                <input asp-for="@Model.Driver.Email" class="form-control" />
-                <span asp-validation-for="@Model.Driver.Email" class="text-danger"></span>
-            </div>
-
-            <div class="form-group">
-                <label asp-for="@Model.Driver.PhoneNumber" class="control-label"></label>
-                <input asp-for="@Model.Driver.PhoneNumber" class="form-control" />
-                <span asp-validation-for="@Model.Driver.PhoneNumber" class="text-danger"></span>
-            </div>
-
-            <div class="form-group">
-                <label asp-for="@Model.Driver.Password" class="control-label"></label>
-                <input asp-for="@Model.Driver.Password" class="form-control" />
-                <span asp-validation-for="@Model.Driver.Password" class="text-danger"></span>
-            </div>
-
-            <div class="form-group">
-                <label asp-for="@Model.Driver.VehicleInfo" class="control-label"></label>
-                <input asp-for="@Model.Driver.VehicleInfo" class="form-control" />
-                <span asp-validation-for="@Model.Driver.VehicleInfo" class="text-danger"></span>
-            </div>
-
-            <div class="form-group">
-                <label asp-for="@Model.Driver.LicenseNumber" class="control-label"></label>
-                <input asp-for="@Model.Driver.LicenseNumber" class="form-control" />
-                <span asp-validation-for="@Model.Driver.LicenseNumber" class="text-danger"></span>
-            </div>
-
-            <div class="form-group">
-                 <input type="submit" value="Create" class="btn btn-primary" />
-@*                   <a input asp-page="/Index" type="submit" value="Create" class="btn btn-primary">Create Profile</a>
- *@            </div>
-        </form>
-    </div>
-</div>
