--- 2FSemesterProjekt2024/Pages/Drivers/Delete.cshtml
@@ -1,61 +1,0 @@
-﻿@page
-@model _2FSemesterProjekt2024.Pages.Drivers.DeleteModel
-@{
-    ViewData["Title"] = "Delete Driver";
-}
-
-<h1>Delete Driver</h1>
-<h3>Are you sure you want to delete this?</h3>
-<div>
-    <form method="post">
-        <dl class="row">
-            <dt class="col-sm-2">
-                @Html.DisplayNameFor(model => model.Driver.DriverId)
-            </dt>
-            <dd class="col-sm-10">
-                @Html.DisplayFor(model => model.Driver.DriverId)
-            </dd>
-            <dt class="col-sm-2">
-                @Html.DisplayNameFor(model => model.Driver.DriverName)
-            </dt>
-            <dd class="col-sm-10">
-                @Html.DisplayFor(model => model.Driver.DriverName)
-            </dd>
-            <dt class="col-sm-2">
-                @Html.DisplayNameFor(model => model.Driver.Email)
-            </dt>
-            <dd class="col-sm-10">
-                @Html.DisplayFor(model => model.Driver.Email)
-            </dd>
-            <dt class="col-sm-2">
-                @Html.DisplayNameFor(model => model.Driver.PhoneNumber)
-            </dt>
-            <dd class="col-sm-10">
-                @Html.DisplayFor(model => model.Driver.PhoneNumber)
-            </dd>
-            <dt class="col-sm-2">
-                @Html.DisplayNameFor(model => model.Driver.Password)
-            </dt>
-            <dd class="col-sm-10">
-                @Html.DisplayFor(model => model.Driver.Password)
-            </dd>
-            <dt class="col-sm-2">
-                @Html.DisplayNameFor(model => model.Driver.VehicleInfo)
-            </dt>
-            <dd class="col-sm-10">
-                @Html.DisplayFor(model => model.Driver.VehicleInfo)
-            </dd>
-            <dt class="col-sm-2">
-                @Html.DisplayNameFor(model => model.Driver.LicenseNumber)
-            </dt>
-            <dd class="col-sm-10">
-                @Html.DisplayFor(model => model.Driver.LicenseNumber)
-            </dd>
-            
-        </dl>
-        <input type="submit" value="Delete" class="btn btn-danger" />
-    </form>
-    <div>
-        <a asp-page="GetDrivers">Back to List</a>
-    </div>
-</div>
