--- 2FSemesterProjekt2024/Pages/Drivers/Update.cshtml
@@ -1,50 +1,0 @@
-﻿@page 
-@model _2FSemesterProjekt2024.Pages.Drivers.UpdateModel
-@{
-    ViewData["Title"] = "UpdateDriver";
-}
-<h1>Update Driver</h1>
-
-<form method="post">
-    <div class="form-group">
-        <input type="hidden" asp-for="Driver.DriverId" />
-    </div>
-
-    <div class="form-group">
-        <label asp-for="@Model.Driver.DriverName" class="control-label"></label>
-        <input asp-for="@Model.Driver.DriverName" class="form-control" />
-        <span asp-validation-for="@Model.Driver.DriverName" class="text-danger"></span>
-    </div>
-    <div class="form-group">
-        <label asp-for="@Model.Driver.Email" class="control-label"></label>
-        <input asp-for="@Model.Driver.Email" class="form-control" />
-        <span asp-validation-for="@Model.Driver.Email" class="text-danger"></span>
-    </div>
-    <div class="form-group">
-        <label asp-for="@Model.Driver.PhoneNumber" class="control-label"></label>
-        <input asp-for="@Model.Driver.PhoneNumber" class="form-control" />
-        <span asp-validation-for="@Model.Driver.PhoneNumber" class="text-danger"></span>
-    </div>
-    <div class="form-group">
-        <label asp-for="@Model.Driver.Password" class="control-label"></label>
-        <input asp-for="@Model.Driver.Password" class="form-control" />
-        <span asp-validation-for="@Model.Driver.Password" class="text-danger"></span>
-    </div>
-     <div class="form-group">
-            <label asp-for="@Model.Driver.VehicleInfo" class="control-label"></label>
-            <input asp-for="@Model.Driver.VehicleInfo" class="form-control" />
-            <span asp-validation-for="@Model.Driver.VehicleInfo" class="text-danger"></span>
-     </div>
-     <div class="form-group">
-            <label asp-for="@Model.Driver.LicenseNumber" class="control-label"></label>
-            <input asp-for="@Model.Driver.LicenseNumber" class="form-control" />
-            <span asp-validation-for="@Model.Driver.LicenseNumber" class="text-danger"></span>
-     </div>
-        <div class="form-group">
-            <input type="submit" value="Save" class="btn btn-primary" />
-        </div>
-</form>
-
-<div>
-    <a asp-page="GetDriver">Back to List</a>
-</div>
