+++ 2FSemesterProjekt2024/Pages/Error.cshtml
@@ -1,10 +1,10 @@
 ﻿@page
-@model ErrorModel
+@model _2FSemesterProjekt2024.Pages.ErrorModel
 @{
     ViewData["Title"] = "Error";
 }
 
-<h1 class="text-danger">Error.</h1>
+<h1 class="text-danger">Error</h1>
 <h2 class="text-danger">An error occurred while processing your request.</h2>
 
 @if (Model.ShowRequestId)
@@ -14,13 +14,23 @@
     </p>
 }
 
-<h3>Development Mode</h3>
-<p>
-    Swapping to the <strong>Development</strong> environment displays detailed information about the error that occurred.
-</p>
-<p>
-    <strong>The Development environment shouldn't be enabled for deployed applications.</strong>
-    It can result in displaying sensitive information from exceptions to end users.
-    For local debugging, enable the <strong>Development</strong> environment by setting the <strong>ASPNETCORE_ENVIRONMENT</strong> environment variable to <strong>Development</strong>
-    and restarting the app.
-</p>
+@if (Environment.GetEnvironmentVariable("ASPNETCORE_ENVIRONMENT") == "Development")
+{
+    <h3>Development Mode</h3>
+    <p>
+        Detailed information about the error is shown when the application is in <strong>Development</strong> mode.
+    </p>
+    <p>
+        <strong>Warning:</strong> The Development environment shouldn't be enabled for deployed applications, as it may expose sensitive information.
+    </p>
+    <p>
+        To enable development mode for local debugging, set the <strong>ASPNETCORE_ENVIRONMENT</strong> environment variable to <strong>Development</strong> and restart the app.
+    </p>
+}
+else
+{
+    <h3>Production Mode</h3>
+    <p>
+        We apologize, but an unexpected error occurred. Please try again later, or contact support if the issue persists.
+    </p>
+}
