+++ 2FSemesterProjekt2024/Pages/Index.cshtml
@@ -1,5 +1,5 @@
 ﻿@page
-@model IndexModel
+@model _2FSemesterProjekt2024.Pages.IndexModel
 @{
     ViewData["Title"] = "Home page";
 }
