--- 2FSemesterProjekt2024/Pages/Passengers/Create.cshtml
@@ -1,41 +1,0 @@
-﻿@page
-@model _2FSemesterProjekt2024.Pages.Passengers.CreateModel
-@{
-    ViewData["Title"] = "Create";
-}
-<h1>Create</h1>
-<div class="row">
-    <div class="col-md-4">
-        <form method="post">
-            <div asp-validation-summary="  ModelOnly  " class="text-danger">
-            </div>
-
-            <div class="form-group">
-                <label asp-for="@Model.Passenger.PassengerName" class="control-label"></label>
-                <input asp-for="@Model.Passenger.PassengerName" class="form-control" />
-                <span asp-validation-for="@Model.Passenger.PassengerName" class="text-danger"></span>
-            </div>
-            <div class="form-group">
-                <label asp-for="@Model.Passenger.Email" class="control-label"></label>
-                <input asp-for="@Model.Passenger.Email" class="form-control" />
-                <span asp-validation-for="@Model.Passenger.Email" class="text-danger"></span>
-            </div>
-
-            <div class="form-group">
-                <label asp-for="@Model.Passenger.PhoneNumber" class="control-label"></label>
-                <input asp-for="@Model.Passenger.PhoneNumber" class="form-control" />
-                <span asp-validation-for="@Model.Passenger.PhoneNumber" class="text-danger"></span>
-            </div>
-
-            <div class="form-group">
-                <label asp-for="@Model.Passenger.Password" class="control-label"></label>
-                <input asp-for="@Model.Passenger.Password" class="form-control" />
-                <span asp-validation-for="@Model.Passenger.Password" class="text-danger"></span>
-            </div>
-
-            <div class="form-group">
-                <input type="submit" value="Create" class="btn btn-primary" />
-            </div>
-        </form>
-    </div>
-</div>
