--- 2FSemesterProjekt2024/Pages/Passengers/Delete.cshtml
@@ -1,48 +1,0 @@
-﻿@page
-@model _2FSemesterProjekt2024.Pages.Passengers.DeleteModel
-@{
-    ViewData["Title"] = "Delete Passenger";
-}
-<h1>Delete passenger</h1>
-<h3>Are you sure you want to delete this?</h3>
-<div>
-    <form method="post">
-        <dl class="row">
-            <dt class="col-sm-2">
-                @Html.DisplayNameFor(model => model.Passenger.PassengerId)
-            </dt>
-            <dd class="col-sm-10">
-                @Html.DisplayFor(model => model.Passenger.PassengerId)
-            </dd>
-            <dt class="col-sm-2">
-                @Html.DisplayNameFor(model => model.Passenger.PassengerName)
-            </dt>
-            <dd class="col-sm-10">
-                @Html.DisplayFor(model => model.Passenger.PassengerName)
-            </dd>
-            <dt class="col-sm-2">
-                @Html.DisplayNameFor(model => model.Passenger.Email)
-            </dt>
-            <dd class="col-sm-10">
-                @Html.DisplayFor(model => model.Passenger.Email)
-            </dd>
-            <dt class="col-sm-2">
-                @Html.DisplayNameFor(model => model.Passenger.PhoneNumber)
-            </dt>
-            <dd class="col-sm-10">
-                @Html.DisplayFor(model => model.Passenger.PhoneNumber)
-            </dd>
-            <dt class="col-sm-2">
-                @Html.DisplayNameFor(model => model.Passenger.Password)
-            </dt>
-            <dd class="col-sm-10">
-                @Html.DisplayFor(model => model.Passenger.Password)
-            </dd>
-
-        </dl>
-        <input type="submit" value="Delete" class="btn btn-danger" />
-    </form>
-    <div>
-        <a asp-page="GetPassenger">Back to List</a>
-    </div>
-</div>
