+++ 2FSemesterProjekt2024/Pages/Privacy.cshtml
@@ -1,5 +1,5 @@
 ﻿@page
-@model PrivacyModel
+@model _2FSemesterProjekt2024.Pages.PrivacyModel
 @{
     ViewData["Title"] = "Privacy Policy";
 }
