--- 2FSemesterProjekt2024/Pages/Profile/Profile.cshtml
@@ -1,46 +1,0 @@
-﻿@page
-@model _2FSemesterProjekt2024.Pages.Auth.ProfileModel
-@{
-    ViewData["Title"] = "Profile";
-}
-
-<h1>Your Profile</h1>
-
-@if (Model.UserType == "Driver")
-{
-        <div class="driver-profile">
-            <div class="profile-picture">
-                <img src="@Model.Driver.ProfilePictureUrl" alt="Profile Picture" />
-            </div>
-        <h2>Driver Details</h2>
-        <p><strong>Name:</strong> @Model.Driver.DriverName</p>
-        <p><strong>Email:</strong> @Model.Driver.Email</p>
-        <p><strong>Phone Number:</strong> @Model.Driver.PhoneNumber</p>
-        <p><strong>Vehicle Info:</strong> @Model.Driver.VehicleInfo</p>
-        <p><strong>License Number:</strong> @Model.Driver.LicenseNumber</p>
-        <p><strong>Rating:</strong> @Model.Driver.Rating</p>
-    </div>
-    <a class="btn btn-primary" asp-page="/Drivers/Update" asp-route-did="@Model.Driver.DriverId">Edit Profile</a>
-    <a class="btn btn-primary" asp-page="/Drivers/Delete" asp-route-did="@Model.Driver.DriverId">Delete Profile</a>
-}
-else if (Model.UserType == "Passenger")
-{
-        <div class="passenger-profile">
-            <div class="profile-picture">
-                <img src="@Model.Passenger.ProfilePictureUrl" alt="Profile Picture" />
-            </div>
-        <h2>Passenger Details</h2>
-        <p><strong>Name:</strong> @Model.Passenger.PassengerName</p>
-        <p><strong>Email:</strong> @Model.Passenger.Email</p>
-        <p><strong>Phone Number:</strong> @Model.Passenger.PhoneNumber</p>
-    </div>
-    <a class="btn btn-primary" asp-page="/Passengers/Update" asp-route-pid="@Model.Passenger.PassengerId">Edit Profile</a>
-    <a class="btn btn-primary" asp-page="/passengers/Delete" asp-route-pid="@Model.Passenger.PassengerId">Delete Profile</a>
-}
-else
-{
-    <p>Unable to load profile information. Please <a href="/Auth/Login">log in</a> again.</p>
-}
-
-<a href="/Logout/Logout">Logout</a>
-
