--- 2FSemesterProjekt2024/Pages/Shared/_ValidationScriptsPartial.cshtml
@@ -1,2 +1,0 @@
-﻿<script src="~/lib/jquery-validation/dist/jquery.validate.min.js"></script>
-<script src="~/lib/jquery-validation-unobtrusive/jquery.validate.unobtrusive.min.js"></script>
