+++ 2FSemesterProjekt2024/Pages/_ViewImports.cshtml
@@ -1,3 +1,6 @@
-﻿@using _2FSemesterProjekt2024
-@namespace _2FSemesterProjekt2024.Pages
+﻿@using Microsoft.AspNetCore.Identity
+@using _2FSemesterProjekt2024.Areas.Identity
+
+
+_2FSemesterProjekt2024.Pages
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
