--- 2FSemesterProjekt2024/Services/Interfaces/IPassengerService.cs
@@ -1,19 +1,0 @@
-﻿using _2FSemesterProjekt2024.Models;
-using Microsoft.EntityFrameworkCore;
-
-namespace _2FSemesterProjekt2024.Services.Interfaces
-{
-    public interface IPassengerService
-    {
-        public IEnumerable<Passenger> GetPassengers();
-
-        public Passenger GetPassengerById(int pid);
-
-      
-        public void AddPassenger(Passenger passenger);
-
-        public void DeletePassenger(Passenger passenger);
-
-        public void UpdatePassenger(Passenger passenger);
-    }
-}
